refactor: split skill-manage.ts into three focused modules (530→343 lines)
skill-manage.ts exceeded the 350-line hard limit at 530 lines.
Extract two cohesive modules along natural boundaries:

- skill-mount-ops.ts (144 lines): pure filesystem symlink mount/unmount
  operations (mountSkillSymlinks, unmountSkillSymlinks, activeMountTargets)
- skill-query.ts (80 lines): read-only config queries
  (listSkills, querySkill, SkillInfo, SkillDetail)

skill-manage.ts retains: addSkill, removeSkill, cascadeToProjects,
collision guards, config rollback — the CRUD operations that coordinate
config writes with filesystem mutations.

Re-exports preserve backward compatibility for existing consumers.
76/76 plugin tests pass, 21/23 mount-rules tests pass (2 pre-existing).

[宪宪/claude-opus-4-6🐾]
