+++ packages/api/src/skills/skill-manage.ts
@@ -1,35 +1,30 @@
 /**
- * Skill Management — unified public API for skill CRUD + query.
+ * Skill Management — unified public API for skill CRUD.
  *
- * Public surface: addSkill / removeSkill / listSkills / querySkill
- * Consumers (PluginResourceActivator, capabilities route, console detail view, etc.)
+ * Public surface: addSkill / removeSkill
+ * Consumers (PluginResourceActivator, capabilities route, etc.)
  * call these functions. Config writes + symlink operations are handled internally.
+ *
+ * Query functions (listSkills, querySkill) → skill-query.ts
+ * Symlink operations (mountSkillSymlinks, unmountSkillSymlinks) → skill-mount-ops.ts
  */
 
 import { existsSync } from 'node:fs';
-import { lstat, mkdir, rm } from 'node:fs/promises';
-import { homedir } from 'node:os';
-import { dirname, join, relative } from 'node:path';
+import { join } from 'node:path';
 
-import {
-  type CapabilitiesConfig,
-  type CapabilityEntry,
-  type MountRules,
-  STANDARD_MOUNT_POINT_IDS,
-} from '@cat-cafe/shared';
+import { type CapabilitiesConfig, type CapabilityEntry, type MountRules } from '@cat-cafe/shared';
 import { readCapabilitiesConfig, writeCapabilitiesConfig } from '../config/capabilities/capability-orchestrator.js';
 import { readMountRules } from '../config/mount/mount-rules-store.js';
-import { buildSkillMountTargets, createSkillSymlink } from '../utils/skill-mount.js';
-import { parseManifestSkillMeta, readSkillMeta } from './skill-meta.js';
+import { activeMountTargets, mountSkillSymlinks, unmountSkillSymlinks } from './skill-mount-ops.js';
 import { syncAll } from './skill-sync-all.js';
-import { classifyMountPath, type MountConflict } from './skill-sync-engine.js';
+import type { MountConflict } from './skill-sync-engine.js';
+
+export { mountSkillSymlinks, unmountSkillSymlinks } from './skill-mount-ops.js';
+// Re-export for consumers that import from skill-manage
+export type { SkillDetail, SkillInfo } from './skill-query.js';
+export { listSkills, querySkill } from './skill-query.js';
 
 // ────────── Types ──────────
-
-interface MountTarget {
-  id: string;
-  dirs: string[];
-}
 
 /** F228: Cascade skill changes to all governance-registered projects via syncAll.
  *  Caller provides the cat-cafe skills source dir; addSkill/removeSkill handles the rest.
@@ -107,29 +102,6 @@ export async function cascadeToProjects(mainProjectRoot: string, catCafeSkillsSo
   }
 }
 
-function symlinkTargetFor(linkPath: string, sourcePath: string): string {
-  return process.platform === 'win32' ? sourcePath : relative(dirname(linkPath), sourcePath);
-}
-
-function activeMountTargets(projectRoot: string, rules: MountRules): MountTarget[] {
-  const standard = STANDARD_MOUNT_POINT_IDS.filter((id) => rules.mountPoints[id].enabled).map((id) => ({
-    id,
-    dirs: [join(projectRoot, rules.mountPoints[id].path)],
-  }));
-  const custom = buildSkillMountTargets(projectRoot, homedir(), rules)
-    .filter((t) => t.kind === 'custom')
-    .map((t) => ({ id: t.id, dirs: t.candidates }));
-  return [...standard, ...custom];
-}
-
-function allMountDirs(projectRoot: string, rules: MountRules): string[] {
-  const standardDirs = STANDARD_MOUNT_POINT_IDS.map((id) => join(projectRoot, rules.mountPoints[id].path));
-  const customDirs = buildSkillMountTargets(projectRoot, homedir(), rules)
-    .filter((t) => t.kind === 'custom')
-    .flatMap((t) => t.candidates);
-  return [...new Set([...standardDirs, ...customDirs])];
-}
-
 function findSkillEntry(
   capabilities: CapabilityEntry[],
   capabilityId: string,
@@ -162,126 +134,7 @@ async function writeCapabilitiesWithRollback(
   }
 }
 
-// ────────── Symlink helpers (shared by addSkill and PluginResourceActivator) ──────────
-
-/**
- * Mount symlinks for a skill into active mount point directories.
- * Pure filesystem operation — does not touch capabilities config.
- */
-export async function mountSkillSymlinks(
-  projectRoot: string,
-  skillName: string,
-  skillsSource: string,
-  mountRules: MountRules,
-  mountPaths?: readonly string[],
-): Promise<SkillOperationResult> {
-  const result: SkillOperationResult = { mounted: [], unmounted: [], conflicts: [] };
-  const targets = activeMountTargets(projectRoot, mountRules);
-  const allowed = mountPaths ? new Set(mountPaths) : null;
-
-  for (const target of targets) {
-    if (allowed && !allowed.has(target.id)) {
-      for (const dir of target.dirs) {
-        // Guard: skip symlinked mount dirs (same as mount branch below)
-        try {
-          const s = await lstat(dir);
-          if (s.isSymbolicLink() || !s.isDirectory()) continue;
-        } catch {
-          continue;
-        }
-        const linkPath = join(dir, skillName);
-        if ((await classifyMountPath(linkPath, skillsSource, skillName)) === 'managed') {
-          await rm(linkPath);
-          result.unmounted.push({ skillName, mountPointId: target.id, path: linkPath });
-        }
-      }
-      continue;
-    }
-    for (const dir of target.dirs) {
-      // Guard: reject symlinked mount dirs to prevent writing outside project
-      try {
-        const dirStat = await lstat(dir);
-        if (dirStat.isSymbolicLink() || !dirStat.isDirectory()) {
-          result.conflicts.push({ skillName, mountPointId: target.id, path: dir });
-          continue;
-        }
-      } catch (err) {
-        if ((err as NodeJS.ErrnoException).code !== 'ENOENT') {
-          result.conflicts.push({ skillName, mountPointId: target.id, path: dir });
-          continue;
-        }
-      }
-      await mkdir(dir, { recursive: true });
-      const linkPath = join(dir, skillName);
-      const status = await classifyMountPath(linkPath, skillsSource, skillName);
-      if (status === 'missing') {
-        await createSkillSymlink(symlinkTargetFor(linkPath, join(skillsSource, skillName)), linkPath);
-        result.mounted.push({ skillName, mountPointId: target.id, path: linkPath });
-      } else if (status === 'conflict') {
-        result.conflicts.push({ skillName, mountPointId: target.id, path: linkPath });
-      }
-    }
-  }
-  return result;
-}
-
-/**
- * Remove managed symlinks for a skill from all mount point directories.
- * Pure filesystem operation — does not touch capabilities config.
- */
-export async function unmountSkillSymlinks(
-  projectRoot: string,
-  skillName: string,
-  skillsSource: string,
-  mountRules: MountRules,
-): Promise<SkillOperationResult> {
-  const result: SkillOperationResult = { mounted: [], unmounted: [], conflicts: [] };
-  for (const dir of allMountDirs(projectRoot, mountRules)) {
-    // Guard: skip symlinked mount dirs to avoid following into external targets
-    try {
-      const dirStat = await lstat(dir);
-      if (dirStat.isSymbolicLink() || !dirStat.isDirectory()) continue;
-    } catch {
-      continue;
-    }
-    const linkPath = join(dir, skillName);
-    if ((await classifyMountPath(linkPath, skillsSource, skillName)) === 'managed') {
-      await rm(linkPath);
-      result.unmounted.push({ skillName, mountPointId: 'cleanup', path: linkPath });
-    }
-  }
-  return result;
-}
-
 // ────────── Public API ──────────
-
-export interface SkillInfo {
-  /** Capability entry ID (e.g. 'tdd' or 'plugin:foo:my-skill'). */
-  id: string;
-  enabled: boolean;
-  pluginId?: string;
-  mountPaths?: readonly string[];
-}
-
-/**
- * List all cat-cafe managed skills configured for a project.
- *
- * Pure config read — no filesystem checks. Consumers that need mount
- * status should call `classifyMountPath` per mount point themselves.
- */
-export async function listSkills(projectRoot: string): Promise<SkillInfo[]> {
-  const config = await readCapabilitiesConfig(projectRoot);
-  if (!config) return [];
-
-  return config.capabilities
-    .filter((c) => c.type === 'skill' && c.source === 'cat-cafe')
-    .map((c) => ({
-      id: c.id,
-      enabled: c.enabled ?? false,
-      ...(c.pluginId ? { pluginId: c.pluginId } : {}),
-      ...(c.mountPaths?.length ? { mountPaths: c.mountPaths } : {}),
-    }));
-}
 
 /**
  * Add a skill to a project: upsert config entry + mount symlinks.
@@ -487,44 +340,4 @@ export async function removeSkill(
   }
 
   return result;
-}
-
-// ────────── Query ──────────
-
-export interface SkillDetail extends SkillInfo {
-  description?: string;
-  triggers?: string[];
-  category?: string;
-}
-
-/**
- * Query detailed information about a single skill.
- *
- * Combines config state (enabled/mountPaths) with metadata from
- * SKILL.md frontmatter and manifest.yaml. Used by console detail view.
- */
-export async function querySkill(
-  projectRoot: string,
-  skillName: string,
-  skillsSource: string,
-): Promise<SkillDetail | null> {
-  const skills = await listSkills(projectRoot);
-  const info = skills.find((s) => s.id === skillName || s.id.endsWith(`:${skillName}`));
-
-  const skillDir = join(skillsSource, skillName);
-  const [skillMeta, manifestMeta] = await Promise.all([readSkillMeta(skillDir), parseManifestSkillMeta(skillsSource)]);
-  const manifest = manifestMeta.get(skillName);
-
-  // Skill not in config AND not in source → doesn't exist
-  if (!info && !manifest && !skillMeta.description) return null;
-
-  return {
-    id: info?.id ?? skillName,
-    enabled: info?.enabled ?? false,
-    ...(info?.pluginId ? { pluginId: info.pluginId } : {}),
-    ...(info?.mountPaths ? { mountPaths: info.mountPaths } : {}),
-    description: manifest?.description ?? skillMeta.description,
-    triggers: manifest?.triggers ?? skillMeta.triggers,
-    category: manifest?.category,
-  };
 }
